requirements.txt
Cerberus==1.3.4
click==8.1.3
colorama==0.4.5
PyYAML==6.0
requests==2.28.1